chore(credential-providers): move base fromTempCreds impl to non-browser file
